Introduce mvvm
