+++ kitemiru/kitemiru/Kitemiru/Face.swift
@@ -1,0 +1,13 @@
+//
+//  Face.swift
+//  kitemiru
+//
+//  Created by 濵田　悠樹 on 2024/08/17.
+//
+
+import UIKit
+
+public class Face {
+    var types: [FaceType] = FaceType.allCases
+    var ciImages: [CIImage] = []
+}
